trading hub footer logic and implementation
